Change method of generaating Certificate serial numbers #631
When generating Serial numbers from host names a hash os the rootCA,
start_date and end_date is added to the CN to produce a unique serial
number.  This means that the serial number for a host will change if
the rootCA or start/end date is changed.  This will force a re-generation
of the certificate.

The generated cert store should be cleared to remove the now stale
certificates previously generated.